normalize parse result refactor
