+++ src/main/kotlin/astminer/common/model/ParsingModel.kt
@@ -1,5 +1,8 @@
 package astminer.common.model
 
+import astminer.common.preOrder
+import astminer.common.setNormalizedToken
+import astminer.common.splitToSubtokens
 import java.io.File
 import java.io.InputStream
 
@@ -55,4 +58,20 @@ interface Parser<T : Node> {
     }
 }
 
-data class ParseResult<T : Node>(val root: T?, val filePath: String)
+data class ParseResult<T : Node>(val root: T?, val filePath: String) {
+    fun normalize(splitTokens: Boolean) {
+        this.root?.preOrder()?.forEach { node -> astminer.cli.processNodeToken(node, splitTokens) }
+    }
+
+    private fun processNodeToken(node: Node, splitToken: Boolean) {
+        if (splitToken) {
+            node.setNormalizedToken(separateToken(node.getToken()))
+        } else {
+            node.setNormalizedToken()
+        }
+    }
+
+    private fun separateToken(token: String, separator: CharSequence = "|"): String {
+        return splitToSubtokens(token).joinToString(separator)
+    }
+}
