+++ src/main/kotlin/astminer/parse/fuzzy/cpp/FuzzyHandler.kt
@@ -1,5 +1,7 @@
 package astminer.parse
 
+import astminer.common.model.HandlerFactory
+import astminer.common.model.LanguageHandler
 import astminer.common.model.ParseResult
 import astminer.parse.fuzzy.cpp.FuzzyCppParser
 import astminer.parse.fuzzy.cpp.FuzzyMethodSplitter
